requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudscraper==1.2.71
google-search-results==2.4.2
h11==0.14.0
idna==3.10
lxml==5.3.1
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2025.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0